Add checks on uninstall scripts

+++ scripts/uninstall.sh
@@ -1,7 +1,9 @@
 #!/bin/bash
 
-echo "> removing symbolic link for .zshrc"
-rm ~/.zshrc
+if [ -f ~/.zshrc ]; then
+  echo "> removing symbolic link for .zshrc"
+  rm -i ~/.zshrc
+fi
 
 echo "> uninstalling oh-my-zsh"
 sh ./oh-my-zsh/uninstall.sh
